backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
sentence-transformers==3.3.1
scikit-learn==1.5.2
mlxtend==0.23.1
prophet==1.1.6
openai==1.56.1
numpy==1.26.4
pandas==2.2.3
pydantic==2.10.2
torch==2.5.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
